+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Fix predicate argument parsing bugs with commas and parentheses in quoted atoms
 * Fix bug when computing a predicate definition range that would include clauses from other predicates
 * Fix bugs handling diagnostic collections that could result in duplicated diagnostics after fixing and reloading files
+* Fix errors with the debugging commands when the Logtalk terminal unexpectedly doesn't exist
 
 ## [0.63.0]